+++ src/anticarium_desktop/UDPReader.cpp
@@ -17,7 +17,10 @@ UDPReader::UDPReader(QObject* parent) : QObject(parent) {
     // To send data using port defined in settings. Port forwarding for this port must be done
     // in client's router
     const auto& settings = *ApplicationSettings::instance();
-    udp->bind(settings.getClientUDPPort());
+    if(settings.getClientUDPPort())
+    {
+        udp->bind(settings.getClientUDPPort());
+    }
 
     connect(heartbeatTimer, &QTimer::timeout, this, &UDPReader::onHeartbeat);
     connect(reconnectTimer, &QTimer::timeout, heartbeatTimer, qOverload<>(&QTimer::start));